+++ Assets/Scripts/UI/HUD/ComboUI.cs
@@ -1,5 +1,7 @@
+using System;
 using PierreMizzi.Useful;
 using StarWielder.Gameplay.Player;
+using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -13,35 +15,18 @@ namespace StarWielder.UI
 
 		[Header("Main")]
 		[SerializeField] private PlayerChannel m_playerChannel;
-		[SerializeField] private Image m_fillImage;
-		[SerializeField] private int m_maxCombo = 10;
 
-		private float m_normalizedCombo = 0;
-		private float m_fillAmount = 0;
-		private float m_noiseSeed;
-		private float m_noiseValue;
+		[SerializeField] private TextMeshProUGUI m_comboLabel;
 
-		private float NormalizedToFill(float value)
+		private void CallbackComboBreak()
 		{
-			return Mathf.Lerp(m_minFillValue, m_maxFillValue, value);
+			m_comboLabel.text = "";
 		}
 
-		private void CallbackRefreshStarCombo(int combo)
+		private void CallbackComboIncrement()
 		{
-			m_normalizedCombo = combo / (float)m_maxCombo;
+			m_comboLabel.text = m_playerChannel.currentCombo.ToString();
 		}
-
-		#endregion
-
-		#region Settings
-
-		[Header("Settings")]
-		[SerializeField] private Gradient m_gradient;
-
-		[SerializeField] private float m_minFillValue = 0.1f;
-		[SerializeField] private float m_maxFillValue = 0.9f;
-		[SerializeField] private float m_noiseAmplitude = 0.05f;
-		[SerializeField] private float m_noiseFrequency = 15f;
 
 		#endregion
 
@@ -50,16 +35,47 @@ namespace StarWielder.UI
 		private void Start()
 		{
 			if (m_playerChannel != null)
-				m_playerChannel.onRefreshStarCombo += CallbackRefreshStarCombo;
+			{
+				m_playerChannel.onComboIncrement += CallbackComboIncrement;
+				m_playerChannel.onComboBreak += CallbackComboBreak;
+			}
 		}
 
 		private void OnDestroy()
 		{
 			if (m_playerChannel != null)
-				m_playerChannel.onRefreshStarCombo -= CallbackRefreshStarCombo;
+			{
+				m_playerChannel.onComboIncrement -= CallbackComboIncrement;
+				m_playerChannel.onComboBreak -= CallbackComboBreak;
+			}
 		}
 
 		private void Update()
+		{
+			UpdateComboBar();
+		}
+
+		#endregion
+
+		#region Combo Bar
+
+		[Header("Combo Bar")]
+		[SerializeField] private Image m_fillImage;
+		[SerializeField] private Gradient m_gradient;
+
+		[Header("Settings")]
+		[SerializeField] private int m_maxCombo = 10;
+		[SerializeField] private float m_minFillValue = 0.1f;
+		[SerializeField] private float m_maxFillValue = 0.9f;
+		[SerializeField] private float m_noiseAmplitude = 0.05f;
+		[SerializeField] private float m_noiseFrequency = 15f;
+
+		private float m_normalizedCombo => m_playerChannel.currentCombo / (float)m_maxCombo;
+		private float m_fillAmount = 0;
+		private float m_noiseSeed;
+		private float m_noiseValue;
+
+		private void UpdateComboBar()
 		{
 			m_noiseSeed += Time.deltaTime * m_noiseFrequency;
 			m_noiseValue = Mathf.PerlinNoise(m_noiseSeed, 0);
@@ -71,7 +87,14 @@ namespace StarWielder.UI
 			m_fillImage.color = m_gradient.Evaluate(m_fillAmount);
 		}
 
+		private float NormalizedToFill(float value)
+		{
+			return Mathf.Lerp(m_minFillValue, m_maxFillValue, value);
+		}
+
 		#endregion
+
+
 
 	}
 }